# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: クッキングアプリ REST API
  description: |-
    クッキングアプリのREST API
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: https://{ domain }/api/v1
paths:
  /user:
    get:
      tags:
        - User
      summary: ユーザー情報取得API
      description: ユーザーの情報を取得する
      responses:
        '200':
          description: |-
            ユーザー情報をJSONで返す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
    put:
      tags:
        - User
      summary: ユーザー情報編集API
      description: ユーザー情報を編集する
      requestBody:
        description: 編集するユーザー情報
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutUserRequest'
      responses:
        '200':
          description: OK
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
  /user/signup:
    post:
      tags:
        - User
      summary: ユーザー新規登録API
      description: ユーザーを新規登録する
      requestBody:
        description: 新規登録するユーザー情報
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignUpRequest'
      responses:
        '200':
          description: OK
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
  /user/password:
    post:
      tags:
        - User
      summary: ユーザーパスワード変更受付メール配信API
      description: パスワードの変更を行うメールを配信するAPI
      requestBody:
        description: パスワードを変更するユーザーのメールアドレス
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPasswordChangeMailRequest'
      responses:
        '200':
          description: OK
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
    patch:
      tags:
        - User
      summary: ユーザーパスワード変更API
      description: ユーザーのパスワードを変更する
      requestBody:
        description: 新しいパスワード
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPasswordChangeRequest'
      responses:
        '200':
          description: |-
            カテゴリー別に配列形式で返す
            
            一覧表示時のソートは「料理名」の昇順
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
  /dish:
    get:
      tags:
        - Dish
      summary: 料理一覧取得API
      description: 登録された料理の一覧を取得
      parameters:
        - name: category
          in: query
          description: カテゴリ−別に絞り込む
          required: false
          schema:
            type: array
            items:
              type: string
          example: [ JAPAN_FOOD, WESTERN_FOOD ]
          style: form
          explode: true
      responses:
        '200':
          description: |-
            カテゴリー別に配列形式で返す
            
            一覧表示時のソートは「料理名」の昇順
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishesResponse'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
    post:
      tags:
        - Dish
      summary: 料理登録API
      description: 料理登録を行うAPI
      requestBody:
        description: 新規登録する料理情報
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDishRequest'
      responses:
        '200':
          description: |-
            料理idを返す
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                properties:
                  id:
                    type: string
                    description: 料理id
                    example: f59fa544-abfe-423d-a20c-2799eed2d606
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'

  /dish/search:
    get:
      tags:
        - Dish
      summary: 料理検索一覧取得API
      description: サジェスト検索時に使用する一覧取得API
      parameters:
        - name: dishName
          in: query
          description: 入力内容を元に、部分一致で検索する
          required: false
          schema:
            type: string
          example: 季節
      responses:
        '200':
          description: |-
            idとタイトルのJSONを配列で返す
            
            一覧表示時のソートは「料理名」の昇順
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishesSearchResponse'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
  /dish/{dishId}:
    get:
      tags:
        - Dish
      summary: 料理詳細取得API
      description: 登録された料理の一覧を取得
      parameters:
        - name: dishId
          in: path
          description: 料理ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 料理の詳細レスポンスを返す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishDetailResponse'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
    post:
      tags:
        - Dish
      summary: 料理お気に入りAPI
      description: 料理をお気に入り登録するAPI
      parameters:
        - name: dishId
          in: path
          description: 料理ID
          required: true
          schema:
            type: string
      requestBody:
        description: お気に入りフラグ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishFavoriteRequest'
      responses:
        '200':
          description: OK
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
  /dish/{dishId}/process:
    get:
      tags:
        - Dish
      summary: 料理工程取得API
      description: 料理を作る工程を取得するAPI
      parameters:
        - name: dishId
          in: path
          description: 料理ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: |-
            工程の配列と料理名を返す
            
            配列の順番ずつ表示する
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishProcessesResponse'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
  /dish/{dishId}/material:
    get:
      tags:
        - Dish
      summary: 料理材料一覧取得API
      description: 料理の材料一覧を取得するAPI
      parameters:
        - name: dishId
          in: path
          description: 料理ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 材料配列
          content:
            application/json:
              schema:
                type: object
                properties:
                  materials:
                    $ref: '#/components/schemas/DishMaterialResponse'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
  /category:
    get:
      tags:
        - Category
      summary: カテゴリ一覧取得API
      description: 料理のカテゴリ一覧を取得する
      responses:
        '200':
          description: |-
            配列形式で返す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
  /image/upload:
    post:
      tags:
        - Image
      summary: 画像アップロードAPI
      description: 画像をS3にアップロードして、IDを返す
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: |-
            IDを返す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400':
          description: 不正リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: 操作権限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
components:
  schemas:
    UserSignUpRequest:
      type: object
      required:
        - lastName
        - firstName
        - userName
        - email
        - telNumber
      properties:
        userName:
          type: string
          description: ユーザー名
          example: リョーリタロー
        email:
          type: string
          description: email
          example: ryoritaro@hoge.com
        telNumber:
          type: string
          description: 電話番号
          example: "000-0000-0000"
    PutUserRequest:
      type: object
      required:
        - email
        - telNumber
      properties:
        email:
          type: string
          description: Eメール
          example: hoge@hoge.com
        telNumber:
          type: string
          description: 電話番号
          example: "00000000000"
        displayName:
          type: string
          description: 表示用ユーザー名
          example: リョーリタロー
        imageId:
          type: string
          description: 画像アップロードAPIカラ取得した画像ID
          example: f59fa544-abfe-423d-a20c-2799eed2d606
    UserPasswordChangeMailRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          description: Email
          example: ryoritaro@hoge.com
    UserPasswordChangeRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          description: パスワード
          example: "**********"
    UserResponse:
      type: object
      required:
        - name
        - image
        - email
        - telNumber
      properties:
        name:
          type: string
          description: ユーザー名
          example: リョーリタロー
        image:
          type: string
          description: 画像
          example: https://image.png
        email:
          type: string
          description: Email
          example: ryoritaro@hoge.com
        telNumber:
          type: string
          description: 電話番号
          example: 000-0000-0000
        nickname:
          type: string
          description: 表示用ユーザ名（ニックネーム）
          example: タロー
    DishFavoriteRequest:
      type: object
      required:
        - favoriteFlag
      properties:
        favoriteFlag:
          description: お気に入りフラグ
          type: boolean
          example: true
    DishesResponse:
      type: object
      properties:
        dishes:
          description: 料理一覧
          type: array
          items:
            $ref: '#/components/schemas/DishResponse'
    DishResponse:
      type: object
      required:
        - dishId
        - dishName
        - dishCreateRequiredTime
      properties:
        dishId:
          description: 料理ID
          type: string
          example: f59fa544-abfe-423d-a20c-2799eed2d606
        dishName:
          description: 料理名
          type: string
          example: 季節の野菜たっぷりのカレー
        image:
          description: 画像
          type: string
          example: https://image.png
        dishCreateRequiredTime:
          description: 所要時間（分）
          type: number
          example: 20
    PostDishRequest:
      type: object
      required:
        - dishName
        - createRequiredTime
        - imageIds
        - materials
        - category
      properties:
        dishName:
          type: string
          description: 料理名
          example: お好み焼き
        createRequiredTime:
          type: number
          description: 所要時間
          example: 20
        imageIds:
          description: 画像アップロードAPIで取得したid配列
          type: array
          items:
            type: string
            example: f59fa544-abfe-423d-a20c-2799eed2d606
        materials:
          type: array
          items:
            $ref: '#/components/schemas/PostMaterialRequest'
        category:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    DishesSearchResponse:
      type: object
      properties:
        dishesSearch:
          description: 料理検索一覧
          type: array
          items:
            $ref: '#/components/schemas/DishSearchResponse'
    DishSearchResponse:
      type: object
      required:
        - dishId
        - dishName
      properties:
        dishId:
          description: ID
          type: string
          example: f59fa544-abfe-423d-a20c-2799eed2d606
        dishName:
          description: 料理名
          type: string
          example: 季節の野菜たっぷりのカレー
    DishDetailResponse:
      type: object
      required:
        - id
        - nane
        - images
        - categories
        - createRequiredTime
        - favoriteFlag
      properties:
        id:
          description: ID
          type: string
          example: f59fa544-abfe-423d-a20c-2799eed2d606
        name:
          description: 料理名
          type: string
          example: 季節の野菜たっぷりのカレー
        images:
          description: 画像配列
          type: array
          items:
            type: object
            description: 画像
            properties:
              id:
                type: string
                description: 料理画像ID
                example: f59fa544-abfe-423d-a20c-2799eed2d606
              url:
                type: string
                description: 料理画像パス
                example: https://image.png
        categories:
          description: カテゴリー
          type: array
          items:
            $ref: '#/components/schemas/Category'
        createRequiredTime:
          description: 所要時間（分）
          type: number
          example: 20
        favoriteFlag:
          description: お気に入りフラグ
          type: boolean
          example: true
    DishProcessResponse:
      type: object
      description: 工程
      properties:
        dishProcessId:
          description: ID
          type: string
          example: f59fa544-abfe-423d-a20c-2799eed2d606
        dishProcessText:
          description: 説明文
          type: string
          example: にんじんをイチョウ切りにします
    DishProcessesResponse:
      type: object
      description: 工程
      properties:
        processes:
          type: array
          description: 工程配列
          items:
            $ref: '#/components/schemas/DishProcessResponse'
        dishName:
          description: 料理名
          type: string
          example: 季節の野菜たっぷりのカレー
    DishMaterialResponse:
      type: array
      description: 材料配列
      items:
        $ref: '#/components/schemas/MaterialResponse'
    MaterialResponse:
      type: object
      description: 材料
      properties:
        id:
          type: string
          description: 材料ID
          example: f59fa544-abfe-423d-a20c-2799eed2d606
        name:
          type: string
          description: 材料名
          example: にんじん
        quantity:
          type: number
          description: 個数
          example: 2
        unit:
          type: string
          description: 単位
          example: 大さじ
    PostMaterialRequest:
      type: object
      description: 材料
      properties:
        materialId:
          type: string
          description: 材料ID
          example: f59fa544-abfe-423d-a20c-2799eed2d606
        materialName:
          type: string
          description: 材料名
          example: にんじん
        quantity:
          type: number
          description: 個数
          example: 2
        unit:
          type: string
          description: 単位
          example: UNIT
    CategoryResponse:
      $ref: '#/components/schemas/Categories'
    Category:
      type: string
      description: |-
        - JAPAN_FOOD - 和食
        
        - WESTERN_FOOD - 洋食
        
        - CHINESE_FOOD - 中華
        
        - MEAT_DISH - 肉料理
        
        - FISH_DISH - 魚料理
        
        - NOODLE - 麺
        
        - RICE - 米
        
        - SALAD - サラダ
        
        - WITH_ALCOHOL - お酒に合う
      enum:
        - JAPAN_FOOD
        - WESTERN_FOOD
        - CHINESE_FOOD
        - MEAT_DISH
        - FISH_DISH
        - NOODLE
        - RICE
        - SALAD
        - WITH_ALCOHOL
    Categories:
      type: object
      properties:
        categories:
          description: カテゴリー配列
          type: array
          items:
            $ref: '#/components/schemas/Category'
    ImageResponse:
      type: object
      properties:
        id:
          description: 画像ID
          type: string
          example: f59fa544-abfe-423d-a20c-2799eed2d606
    BadRequestException:
      type: object
      required:
        - message
        - httpStatus
        - httpMethod
      properties:
        message:
          description: エラーメッセージ
          type: string
          example: 不正なリクエストです
        httpStatus:
          description: HTTPステータス
          type: number
          example: 400
        httpMethod:
          description: HTTPメソッド
          type: string
          example: GET
        fieldErrors:
          description: フィールドエラー
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    Unauthorized:
      type: string
      example: Unauthorized
    Forbidden:
      type: string
      example: Forbidden
    FieldError:
      type: object
      properties:
        fieldName:
          description: フィールド名
          type: string
          example: userName
        message:
          description: フィールドメッセージ
          type: string
          example: 必須項目です